Fix rgba

+++ runtime/src/altv-c-api/rgba.h
@@ -1,8 +1,8 @@
 #pragma once
 
 typedef struct {
-    float r;
-    float g;
-    float b;
-    float a;
+    uint8_t r;
+    uint8_t g;
+    uint8_t b;
+    uint8_t a;
 } rgba_t;
